api/requirements.txt
Flask==3.1.0
torch==2.5.1 --extra-index-url https://download.pytorch.org/whl/cpu
pythainlp==5.0.5
tensorflow-cpu==2.18.0
gunicorn==23.0.0
scikit-learn==1.6.0
flask-cors==3.0.10